增加语言选择以及语言配置错误自动修正功能; setting.php => settings.php
